Move Mergeable to Util package
