+++ presto-main/src/main/java/com/facebook/presto/operator/exchange/LocalExchangeBufferInfo.java
@@ -13,8 +13,8 @@
  */
 package com.facebook.presto.operator.exchange;
 
-import com.facebook.presto.operator.Mergeable;
 import com.facebook.presto.operator.OperatorInfo;
+import com.facebook.presto.util.Mergeable;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
